frontend: refactor settings cluster
